feat: AppDelegate wires FullDesktopWindow swap logic and providers
- Add fullDesktopWindow, fullDesktopViewModel, metricsProvider properties
- Start/stop providers in applicationDidFinishLaunching/applicationWillTerminate
- Add openFullDesktopMode() / closeFullDesktopMode() with 150ms crossfade
- Observe hudViewModel.$windowMode to drive HUD↔FullDesktop swap
- Instantiate FullDesktopWindow in setupHUD()
- Add @MainActor annotation to AppDelegate (required for @MainActor FullDesktopViewModel/SystemMetricsProvider stored properties)
- Register FullDesktopView, FullDesktopViewModel, FullDesktopWindow, SystemMetricsProvider in Xcode project
- Fix Int vs Int64 type mismatch in SystemMetricsProvider.diskUsage()

+++ jarvis-swift/Jarvis/SystemMetricsProvider.swift
@@ -99,7 +99,7 @@ final class SystemMetricsProvider: ObservableObject {
         let total = values.volumeTotalCapacity,
         let available = values.volumeAvailableCapacityForImportantUsage,
         total > 0 else { return diskPercent }
-        let used = total - available
+        let used = Int64(total) - available
         return min(Double(used) / Double(total) * 100, 100)
     }
